Split up adding kernel and adding initramfs.
In terms of the transaction running this stuff, we want the stanza to
get added with kernel parameters during package additions, so that we
can reap arguments from the kernels being replaced.  At the same time,
the initramfs should be generated and added after the transaction has
completed, so that the initramfs is built with the new userland instead
of some (plausibly completely broken) intermediate state.

Signed-off-by: Peter Jones <pjones@redhat.com>